Added an Auth0 class which helps testing
