+++ .gitignore
@@ -1,1 +1,2 @@
 cuda-repo-wsl-ubuntu-12-4-local_12.4.1-1_amd64.deb
+.env
